+++ app/Filament/App/Clusters/LocationLevels/Resources/FarmResource.php
@@ -34,7 +34,7 @@ class FarmResource extends Resource
             ->schema([
 
                 Forms\Components\Select::make('location_id')
-                    ->label('Select the '.$locationLevelWithFarms->name.' for this farm')
+                    ->label('Select the ' . $locationLevelWithFarms->name . ' for this farm')
                     ->options($locationLevelWithFarms->locations->pluck('name', 'id')),
 
                 Forms\Components\TextInput::make('team_code')
@@ -83,26 +83,26 @@ class FarmResource extends Resource
     {
         $farms = Farm::all()->where('owner_id', HelperService::getCurrentOwner()->id);
 
-        $locationLevelColumns = $farms->map(fn (Farm $farm) => $farm->location->locationLevel)
+        $locationLevelColumns = $farms->map(fn(Farm $farm) => $farm->location->locationLevel)
             ->unique()
             ->values()
             ->map(
-                fn (LocationLevel $locationLevel) => Tables\Columns\TextColumn::make("location_{$locationLevel->id}")
-                    ->getStateUsing(fn ($record) => $record->location->location_level_id === $locationLevel->id ? $record->location->name : '')
+                fn(LocationLevel $locationLevel) => Tables\Columns\TextColumn::make("location_{$locationLevel->id}")
+                    ->getStateUsing(fn($record) => $record->location->location_level_id === $locationLevel->id ? $record->location->name : '')
                     ->label($locationLevel->name)
                     ->sortable()
                     ->searchable()
             );
 
-        $identifiers = $farms->map(fn (Farm $farm) => $farm->identifiers?->keys())
+        $identifiers = $farms->map(fn(Farm $farm) => $farm->identifiers?->keys())
             ->flatten()->unique()->values();
 
-        $idColumns = $identifiers->map(fn ($identifier) => Tables\Columns\TextColumn::make("identifiers.{$identifier}")->label(ucfirst($identifier))->sortable()->searchable());
+        $idColumns = $identifiers->map(fn($identifier) => Tables\Columns\TextColumn::make("identifiers.{$identifier}")->label(ucfirst($identifier))->sortable()->searchable());
 
-        $properties = $farms->map(fn (Farm $farm) => $farm->properties?->keys())
+        $properties = $farms->map(fn(Farm $farm) => $farm->properties?->keys())
             ->flatten()->unique()->values();
 
-        $propertyColumns = $properties->map(fn ($property) => Tables\Columns\TextColumn::make("properties.{$property}")->label(ucfirst($property))->sortable()->searchable());
+        $propertyColumns = $properties->map(fn($property) => Tables\Columns\TextColumn::make("properties.{$property}")->label(ucfirst($property))->sortable()->searchable());
 
         return $table
             ->columns([
@@ -137,6 +137,7 @@ class FarmResource extends Resource
             'index' => Pages\ListFarms::route('/'),
             'create' => Pages\CreateFarm::route('/create'),
             'edit' => Pages\EditFarm::route('/{record}/edit'),
+            'import' => Pages\ImportLocationsAndFarms::route('/import'),
         ];
     }
 }
